// metal_angular/src/app/models/customer.model.ts
export class Customer{
  // @ts-ignore
  id?: number;
  ledger_name: string;
  billing_name: string;
  ledger_group_id: number;
  customer_category_id: number;
  email: string;
  mobile1: string;
  mobile2: string;
  address1: string;
  address2: string;
  state_id: number;
  po: string;
  area: string;
  city: string;
  pin: number;
  transaction_type_id: number;
  opening_balance: number;

}